+++ tests/test_mappings.py
@@ -1,7 +1,40 @@
 from collections.abc import ItemsView, KeysView, ValuesView
 
+from pytest import raises
+
 from pipe_utils.mappings import *
 from pipe_utils.utils import fdiv_by, mul_by
+from pipe_utils.values import gt, is_even
+
+
+def test_filter_keys():
+    assert filter_keys(is_even)({}) == {}
+    data = {1: "a", 2: "b", 3: "c", 4: "d"}
+    assert filter_keys(is_even)(data) == {2: "b", 4: "d"}
+    assert filter_keys(gt(0))(data) == data
+
+
+def test_filter_values():
+    assert filter_values(is_even)({}) == {}
+    data = {"a": 1, "b": 2, "c": 3, "d": 4}
+    assert filter_values(is_even)(data) == {"b": 2, "d": 4}
+    assert filter_values(gt(0))(data) == data
+
+
+def test_get_value():
+    assert get_value(1)({1: "a"}) == "a"
+
+    with raises(KeyError):
+        assert get_value(2)({1: "a"})
+    with raises(KeyError):
+        assert get_value(2)({})
+
+
+def test_get_value_or_default():
+    assert get_value_or_default(1, None)({1: "a"}) == "a"
+    assert get_value_or_default(2, None)({1: "a"}) is None
+    assert get_value_or_default(2, 0)({}) == 0
+    assert get_value_or_default(-1, 0)({}) == 0
 
 
 def test_item_view():
@@ -28,8 +61,9 @@ def test_map_keys():
 def test_map_values():
     data = {1: "a", 2: "b", 3: "c"}
     assert dict(map_values(mul_by(2))(data)) == {1: "aa", 2: "bb", 3: "cc"}
-    data = {'a': 2, 'd': 3, 'b': 4, 'e': 5}
-    assert dict(map_values(fdiv_by(2))(data)) == {'a': 1, 'd': 1, 'b': 2, 'e': 2}
+    data = {"a": 2, "d": 3, "b": 4, "e": 5}
+    expect = {"a": 1, "d": 1, "b": 2, "e": 2}
+    assert dict(map_values(fdiv_by(2))(data)) == expect
 
 
 def test_sorted_by_key():