+++ CHANGELOG.md
@@ -6,7 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ### Additions
 
-- `map_keys` and `map_values` functions in the `mappings` module
-- `utils` module that imports all utility functions
+- `to_each` in the `iterables` module
+- `clamp`, `lclamp`, and `rclamp` in the `values` module
+- `map_keys`, `map_values`, `filter_keys`, `filter_values`, `get_value`,
+  `get_value_or_default` functions in the `mappings` module
+- `utils` module that imports `Pipe`, `Then` and all utility functions
 
-[0.0.2]: https://github.com/James-Ansley/pipe-utils/compare/v0.0.1...v0.1.0
+[0.1.0]: https://github.com/James-Ansley/pipe-utils/compare/v0.0.1...v0.1.0